Rename dataplane middleware
